move layout.ts tests inline
